Generate a waffle chart of journal paper percentages, by continent (each square = 1% of data)

Generate a waffle chart of journal paper percentages, by continent (each square = 1% of data)

Usage

waffle_country_journal(
  data,
  citation = NULL,
  citation_size = NULL,
  journal_abbreviation = TRUE,
  data_formatted = NULL
)

Arguments

data

The processed dataframe of data

citation

Optionally, a citation to add as a footer.

citation_size

Font size of the citation.

journal_abbreviation

Logical, whether to use the journal abbreviation to fit the entire plot, otherwise some journal names can be quite long and accordingly be cropped.

data_formatted

To save time, it is possible to reuse the dataframe used for this function that is generated by table_country_journal().

Examples

if (FALSE) { # \dontrun{
data <- fetch_openalex_pubs(journal_name = "Collabra", pages = 1)
data <- clean_journals_continents(data)
waffle_country_journal(data)
} # }